+++ tests/difference/core/driver/setup.go
@@ -22,7 +22,8 @@ import (
 
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	ibctestingutils "github.com/cosmos/interchain-security/testutil/ibc_testing_utils"
+	ibctestingutils "github.com/cosmos/interchain-security/testutil/ibc_testing"
+	simibc "github.com/cosmos/interchain-security/testutil/simibc"
 
 	cryptoEd25519 "crypto/ed25519"
 
@@ -48,7 +49,7 @@ import (
 
 type Builder struct {
 	suite          *suite.Suite
-	link           ibctestingutils.OrderedLink
+	link           simibc.OrderedLink
 	path           *ibctesting.Path
 	coordinator    *ibctesting.Coordinator
 	clientHeaders  map[string][]*ibctmtypes.Header
@@ -484,7 +485,7 @@ func (b *Builder) createConsumerGenesis(tmConfig *ibctesting.TendermintConfig) *
 }
 
 func (b *Builder) createLink() {
-	b.link = ibctestingutils.MakeOrderedLink()
+	b.link = simibc.MakeOrderedLink()
 	// init utility data structures
 	b.mustBeginBlock = map[string]bool{P: true, C: true}
 	b.clientHeaders = map[string][]*ibctmtypes.Header{}
@@ -567,7 +568,7 @@ func (b *Builder) sendEmptyVSCPacketToFinishHandshake() {
 
 	b.updateClient(b.chainID(C))
 
-	ack, err := ibctestingutils.TryRecvPacket(b.endpoint(P), b.endpoint(C), packet)
+	ack, err := simibc.TryRecvPacket(b.endpoint(P), b.endpoint(C), packet)
 
 	b.link.AddAck(b.chainID(C), ack, packet)
 
@@ -599,7 +600,7 @@ func (b *Builder) beginBlock(chainID string) {
 
 func (b *Builder) updateClient(chainID string) {
 	for _, header := range b.clientHeaders[b.otherID(chainID)] {
-		err := ibctestingutils.UpdateReceiverClient(b.endpointFromID(b.otherID(chainID)), b.endpointFromID(chainID), header)
+		err := simibc.UpdateReceiverClient(b.endpointFromID(b.otherID(chainID)), b.endpointFromID(chainID), header)
 		if err != nil {
 			b.coordinator.Fatal("updateClient")
 		}
@@ -612,7 +613,7 @@ func (b *Builder) deliver(chainID string) {
 	for _, p := range packets {
 		receiver := b.endpointFromID(chainID)
 		sender := receiver.Counterparty
-		ack, err := ibctestingutils.TryRecvPacket(sender, receiver, p.Packet)
+		ack, err := simibc.TryRecvPacket(sender, receiver, p.Packet)
 		if err != nil {
 			b.coordinator.Fatal("deliver")
 		}
@@ -622,7 +623,7 @@ func (b *Builder) deliver(chainID string) {
 
 func (b *Builder) deliverAcks(chainID string) {
 	for _, ack := range b.link.ConsumeAcks(b.otherID(chainID), 999999) {
-		err := ibctestingutils.TryRecvAck(b.endpointFromID(b.otherID(chainID)), b.endpointFromID(chainID), ack.Packet, ack.Ack)
+		err := simibc.TryRecvAck(b.endpointFromID(b.otherID(chainID)), b.endpointFromID(chainID), ack.Packet, ack.Ack)
 		if err != nil {
 			b.coordinator.Fatal("deliverAcks")
 		}
