+++ testutil/simibc/relay_util.go
@@ -1,4 +1,4 @@
-package ibc_testing_utils
+package simibc
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
